Cleanup scala recipe
