+++ couchbase-persistence/couchbase-persistence-scala-sbt/hello-impl/src/main/resources/application.conf
@@ -1,23 +1,13 @@
-#
-#
 play.application.loader = org.example.hello.impl.HelloLoader
 play.allowGlobalApplication = true
 
-# The properties below override Lagom default configuration with the recommended values for new projects.
-#
-# Lagom has not yet made these settings the defaults for backward-compatibility reasons.
-
-# Prefer 'ddata' over 'persistence' to share cluster sharding state for new projects.
-# See https://doc.akka.io/docs/akka/current/cluster-sharding.html#distributed-data-vs-persistence-mode
 akka.cluster.sharding.state-store-mode = ddata
-
-# Enable the serializer provided in Akka 2.5.8+ for akka.Done and other internal
-# messages to avoid the use of Java serialization.
 akka.actor.serialization-bindings {
   "akka.Done" = akka-misc
   "akka.actor.Address" = akka-misc
   "akka.remote.UniqueAddress" = akka-misc
 }
+
 
 //#couchbase-begin
 hello.couchbase.bucket = "akka"
@@ -30,7 +20,6 @@ hello.couchbase.connection {
 couchbase-journal {
   write.bucket = ${hello.couchbase.bucket}
   snapshot.bucket = ${hello.couchbase.bucket}
-
   # connection settings shared by write, read and snapshot
   connection = ${hello.couchbase.connection}
 }
@@ -38,8 +27,6 @@ couchbase-journal {
 lagom.persistence.read-side.couchbase {
   # bucket must be provided by application.conf
   bucket = ${hello.couchbase.bucket}
-
   connection = ${hello.couchbase.connection}
-
 }
 //#couchbase-end
